Fix consumer name length validator
This patch fixes a mismatch between the size of the column for a
consumer "name" in the database and the value being checked by the api
validator.

The maximum size in the database is 36 chars [1], so we must use that value
in the validator.

[1] https://opendev.org/openstack/barbican/src/branch/stable/xena/barbican/model/models.py#L826

Story: 2009672
Task: 43934

Change-Id: I76f075a94056aa65cd44fd1d7f5d4b24109b6ed1
(cherry picked from commit 31aa926175b62b347593acb8b147810b4b49b9de)