Temporarily disable voting for FIPS job
The gate job barbican-tox-functional-fips is failing at the gate due to
a dependency issue when building the environment.  Specifically, it
appears that the package "liberasurecode-devel" fails to be found in the
CentOS 9 repositories.

This patch temporarily disables gate-voting for the FIPS job.  We should
be able to re-enable voting once this dependency issue is solved.

Change-Id: I9d8028454468f95bae405677dcd492fa2e52f93f