zuul: fix the grenade job to actually test barbican
Add the required services and run a few barbican-specific tests
to validate the upgrade.

The grenade plugin contains a few settings which don't need to be
set anymore explicitly and they are not in the job configuration
(as devstack/upgrade/settings is not used anymore):
- all the image-related variables don't need to be overridden anymore,
  the default one from devstack should be used
- Image API v1 has been disabled since tempest 20.0

The job can be switched to voting again.

Change-Id: Id0682aea57d4d1fc49334f2dd11ef9a0ffb355fb