Fix remaining Secure RBAC policies
This patch fixes the remaining default policies to ensure they are only
evaluated when enforce_new_defaults=True.

The 'reader' role is removed from the policy tests in thise codebase
because they now correctly fail.

The 'reader' role in was introduced into the policy tests in
060ca2ee369af1fff3bf833d466f33d1889fb72e and the dynamic tests that were
generated passed because of the issue being resolved in this patch.

Now that the new default policies are not being evaluated when
enforce_new_defaults=False the requests using the 'reader' role are
correctly denied.

Story: 2010235
Task: 46036
Change-Id: Ief495c50bb120e2aa671dbcc80734ccb5a839b74