db: Replace use of reverse cascades
Resolve the following RemovedIn20Warning warning:

  "SecretStoreMetadatum" object is being merged into a Session along the
  backref cascade path for relationship "Secret.secret_store_metadata";
  in SQLAlchemy 2.0, this reverse cascade will not take place.  Set
  cascade_backrefs to False in either the relationship() or backref()
  function for the 2.0 behavior; or to set globally for the whole
  Session, set the future=True flag

In effect, this means if you have a model that refers to another model,
creating/saving the former will no longer create/save the latter. We
have only one instance of this error - the error message above - and in
our case we are explicitly saving the 'Secret' instance before saving
the 'SecretStoreMetadatum' instance. As such, we can opt-in to the 2.0
behavior with no further changes. We do this for all relationships to be
safe.

More information on this issue can be found at [1].

[1] https://groups.google.com/g/sqlalchemy/c/VoY-qEiJA3U?pli=1

Change-Id: I4b4fa4c224113863643e16153478183447796146
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>