Update gate jobs as per the 2025.1 cycle testing runtime
As per 2025.1 testing runtime[1], we need to test on Ubuntu
Noble (which will be taken care by depends-on tempest and devstack
patches to move base jobs to Noble) and at least single job to run on
Ubuntu Jammy (for smooth upgrade from previous releases).

This commit adds a new job to run on Jammy which can be removed
in future cycle when testing runtime test next version of Ubuntu
as default.

Currently doc job running on Noble fail with the below error
which is fixed by installing libjpeg-dev package
 - ERROR: ERROR: Failed to build installable wheels for some pyproject.toml based projects (Pillow)

Closes-Bug: #2088355

Depends-On: https://review.opendev.org/c/openstack/barbican-tempest-plugin/+/932946

[1] https://governance.openstack.org/tc/reference/runtimes/2025.1.html

Change-Id: I1b6332763c02b08cf2d1e0be9f1ed46e73b963f0