Use MultiFernet to enable more than one KEK
This patch changes the `kek` option for the Simple Crypto Plugin to
allow more than one KEK to be specified.

When more than one KEK is configured, the first KEK is used to encrypt
new data and any additiona KEKs are only used to decrypt existing data.

This change allows for rotating in new KEKs on demand.

Change-Id: I0c3683e316e78478461f5f30f4f353ff43a3bb09