Fix href construction when base url ends with slash
Under certain deployment conditions the `pecan.request.application_url`
may end with the slash (perhaps when wsgi SCRIPT_NAME is set to "/").
And so the url may look like `http://API.DOMAIN//v1/secrets/HASH`. In
any way it is reasonable to ensure base does not end with "/" before we
use it for concatenation.

Change-Id: Ida6a831baff127212fed14958e464847edb5e1b8