testing: Increase HTTP request timeout to 60 seconds
We have been seeing frequent intermittent failures in the
barbican-tox-functional CI job due to timeouts waiting for HTTP
requests to respond, for example:

  requests.exceptions.ReadTimeout:
    HTTPConnectionPool(host='213.32.72.107', port=80): Read timed out.
    (read timeout=10)

This occurred from a request to Barbican API to delete a secret.

The timeout is defaulted to 10 seconds, which does not seem to be long
enough for the current CI node environment. This increases the timeout
to 60 seconds based on the default HTTP request timeout in Tempest [1]
in an effort to improve stability of the barbican-tox-functional job.

[1] https://github.com/openstack/tempest/blob/fc2868e4752204cf470150a37a7d6814d78193d5/tempest/config.py#L229-L232

Change-Id: I2645073ca1fd5085fd165dc1dba4766d77a692a3